Make database generic over `Store` trait.
Any object that implements the `Store` trait can be stored by the
database. The trait itself is generic over the error that could be
raised when serializing the object.

To this end, the derive_more crate was added as a dependency to help
with deriving the `Display` impls for implementing `Error`